% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{bayesian_visual}
\alias{bayesian_visual}
\title{Bayesian Results Visual}
\usage{
bayesian_visual(posterior_control, posterior_treatment, n_sims = 10000)
}
\arguments{
\item{posterior_control}{List with alpha and beta for control posterior}

\item{posterior_treatment}{List with alpha and beta for treatment posterior}

\item{n_sims}{Number of simulated draws (default 10000)}
}
\description{
Plots the posterior distribution of the lift (treatment - control).
}
