% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{both_visual}
\alias{both_visual}
\title{Side-by-Side Frequentist and Bayesian Visuals}
\usage{
both_visual(
  freq_results,
  posterior_control,
  posterior_treatment,
  n_sims = 10000
)
}
\arguments{
\item{freq_results}{Output from ab_freq_test}

\item{posterior_control}{List with alpha and beta for control posterior}

\item{posterior_treatment}{List with alpha and beta for treatment posterior}

\item{n_sims}{Number of simulated draws (default 10000)}
}
\description{
Side-by-Side Frequentist and Bayesian Visuals
}
