% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_data}
\alias{experiment_data}
\title{Experiment Data}
\format{
A dataframe with 10,000 rows and 2 columns:
\describe{
\item{test.group}{Treatment group: "ad" or "psa"}
\item{converted}{Outcome of 1 or 0}}
}
\source{
\url{https://www.kaggle.com/datasets/faviovaz/marketing-ab-testing}
}
\usage{
experiment_data
}
\description{
A sample of 5,000 users per arm from a Kaggle marketing dataset where users
were exposed to either an advertisement or public service announcement, with
conversion recorded as the outcome.
}
\keyword{datasets}
