% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{frequentist_visual}
\alias{frequentist_visual}
\title{Frequentist Results Visual}
\usage{
frequentist_visual(freq_results)
}
\arguments{
\item{freq_results}{Output from ab_freq_test}
}
\description{
Plots the observed effect size with a confidence interval.
}
