% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_data}
\alias{historical_data}
\title{Historical Data}
\format{
A dataframe with 1,000 rows and 2 columns:
\describe{
\item{test.group}{Treatment group: "psa"}
\item{converted}{Outcome of 1 or 0}}
}
\source{
\url{https://www.kaggle.com/datasets/faviovaz/marketing-ab-testing}
}
\usage{
historical_data
}
\description{
The first 1,000 PSA users from the Kaggle marketing dataset, used as
pre-experiment baseline data to inform prior beliefs about conversion rates.
}
\keyword{datasets}
