% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{post_experiment_comparison}
\alias{post_experiment_comparison}
\title{Post-Experiment Analysis Comparison}
\usage{
post_experiment_comparison(
  data,
  outcome_column_name,
  group_column_name,
  control_value,
  treatment_value,
  n,
  alpha = 0.05,
  experiment_prior
)
}
\arguments{
\item{data}{a dataframe of experiment results}

\item{outcome_column_name}{string name of the binary outcome column}

\item{group_column_name}{string name of the column that identifies which
group each observation belongs to}

\item{control_value}{the value in group_column_name that marks control rows}

\item{treatment_value}{the value in group_column_name that marks treatment
rows}

\item{n}{how many observations to use from each group}

\item{alpha}{how often you are willing to declare a winner when there
isn't one}

\item{experiment_prior}{the prior from pre_experiment_comparison, a list
with alpha and beta}
}
\value{
a list containing:
\item{freq_results}{output from ab_freq_test}
\item{bayes_results}{output from bayes_ab_summary}
}
\description{
Runs both frequentist and Bayesian analysis on experiment data. The
frequentist side runs a two-proportion z-test. The Bayesian side updates
the prior with observed data and compares the resulting posteriors.
}
\examples{
data(historical_data)
data(experiment_data)
experiment_prior <- list(alpha = 1, beta = 1, method = "quantile")

post_experiment_data <- post_experiment_comparison(
  data = experiment_data,
  outcome_column_name = "converted",
  group_column_name = "test.group",
  control_value = "psa",
  treatment_value = "ad",
  n = 1000,
  experiment_prior = experiment_prior
)

}
