% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{ab_freq_test}
\alias{ab_freq_test}
\title{Frequentist Proportion Z-test}
\usage{
ab_freq_test(
  conversions_control,
  n_control,
  conversions_treatment,
  n_treatment,
  alpha = 0.05
)
}
\arguments{
\item{conversions_control}{how many conversions occured in the control group}

\item{n_control}{how many observations were in the control group}

\item{conversions_treatment}{how many conversions occured in the test group}

\item{n_treatment}{how many observations were in the test group}

\item{alpha}{how often you are willing to falsely declare a winner when
there isn't one.}
}
\value{
a list containing:
\item{effect}{Observed difference (treatment - control)}
\item{p_value}{Two-sided p-value from the z-test}
\item{ci}{Confidence interval for the difference in proportions}
\item{significant}{Logical. TRUE if p_value < alpha}
}
\description{
This calculates the results of an AB test using a two-proportion z-test.
}
