% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{ab_sample_size}
\alias{ab_sample_size}
\title{Frequentist Sample Size Estimation}
\usage{
ab_sample_size(p_current, mde, alpha = 0.05, power = 0.8)
}
\arguments{
\item{p_current}{current conversion rate. This value should be from 0-1}

\item{mde}{"Minimum detectable effect" this is the smallest change that you
would like to see in the data}

\item{alpha}{how often you are willing to falsely declare a winner when
there isn't one. Lower means you have a stronger test.}

\item{power}{how often you want to correctly detect a real effect when it
exists. higher means you have a stronger test.}
}
\value{
An integer value of how many observations you need PER GROUP
}
\description{
This function will calculate how many observations are needed to detect a
minimum change. It also controls for the error rates that you are willing to
accept.
}
\examples{
ab_sample_size(p_current = 0.02, mde = 0.005)
}
