% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bayes_ab_summary}
\alias{bayes_ab_summary}
\title{Bayesian Proportion Test
Calculates the results of an AB test by analyzing updated or posterior
values of control and treatment results in experiment.}
\usage{
bayes_ab_summary(posterior_control, posterior_treatment, n_sims = 10000)
}
\arguments{
\item{posterior_control}{list containing information for distribution of
control data}

\item{posterior_treatment}{list containing information for distribution of
treatment data}

\item{n_sims}{number of monte carlo draws, or number of simulated
experiments to run. Only drawback of higher values is that it takes longer.}
}
\description{
Bayesian Proportion Test
Calculates the results of an AB test by analyzing updated or posterior
values of control and treatment results in experiment.
}
