% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{elicit_beta_prior}
\alias{elicit_beta_prior}
\title{Form a Beta Prior}
\usage{
elicit_beta_prior(guess, worst_case, best_case)
}
\arguments{
\item{guess}{expected conversion rate}

\item{worst_case}{lower bound of what you think the conversion rate is}

\item{best_case}{upper bound of what you think the conversion rate is}
}
\value{
A list containing:
\item{alpha}{Alpha parameter of the Beta distribution}
\item{beta}{Beta parameter of the Beta distribution}
}
\description{
This takes beliefs about the prior's values and turns them into a real
distribution. If the conversion rate is near .5, we use moment matching, if
it is not, we use quantile matching.
}
