% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{full_comparison}
\alias{full_comparison}
\title{Full A/B Test Comparison}
\usage{
full_comparison(
  freq_results,
  bayes_results,
  posterior_control,
  posterior_treatment,
  alpha = 0.05,
  n_sims = 10000
)
}
\arguments{
\item{freq_results}{Output from ab_freq_test}

\item{bayes_results}{Output from bayes_ab_summary}

\item{posterior_control}{List with alpha and beta for control posterior}

\item{posterior_treatment}{List with alpha and beta for treatment posterior}

\item{alpha}{Significance level (default 0.05)}

\item{n_sims}{Number of simulated draws (default 10000)}
}
\description{
Displays side-by-side visuals and prints interpretation guide.
}
\examples{
freq <- ab_freq_test(
  conversions_control = 100, n_control = 5000,
  conversions_treatment = 150, n_treatment = 5000
)
prior <- elicit_beta_prior(
  guess = 0.02, worst_case = 0.01,
  best_case = 0.03
)
post_ctrl <- update_beta(prior, successes = 100, trials = 5000)
post_trt <- update_beta(prior, successes = 150, trials = 5000)
bayes <- bayes_ab_summary(post_ctrl, post_trt)
full_comparison(freq, bayes, post_ctrl, post_trt)

}
