% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{update_beta}
\alias{update_beta}
\title{Beta Conjugate Update to get posterior}
\usage{
update_beta(prior, successes, trials)
}
\arguments{
\item{prior}{list with alpha and beta from elicit_beta_prior}

\item{successes}{number of observed conversions}

\item{trials}{total number of observations}
}
\description{
Beta Conjugate Update to get posterior
}
